requirements.txt
streamlit>=1.51.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
requests==2.31.0
beautifulsoup4==4.12.2
openpyxl==3.1.2
yfinance>=0.2.66
fredapi==0.5.1
newsapi-python==0.2.7
python-dateutil==2.8.2
python-dotenv==1.0.0
feedparser>=6.0.0
